refactor(server): split value resolution out of properties.rs
properties.rs (extract_properties + extract_property, the pset/property
entity-shape extraction) had grown to 470 lines and tripped the Rust
module-size ratchet gate. Split property VALUE resolution — resolving a
nominal value, enumerated/list members, and IfcComplexProperty nesting
into a display string + kind + measure tag + candidates — into a sibling
property_value.rs. Pure move, no behavior change: extract_properties and
extract_property keep their doc comments and logic verbatim, just import
resolve_single_value/member_list/infer_data_type/fmt_number/
resolve_complex_property_value from the new module.

properties.rs: 470 -> 243 lines. property_value.rs: 237 lines (new).
cargo test -p ifc-lite-server: 275 passed both before and after.
